requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
numpy==1.22.3
openpyxl==3.0.10
pandas==1.4.2
Pillow==9.1.0
setuptools==59.5.0
torch==1.9.0+cu111
torchaudio==0.9.0
torchvision==0.10.0+cu111
tqdm==4.64.0